LayoutTests/fullscreen/full-screen-keyboard-lock-option-enabled.html

<!DOCTYPE html>
<html>
<head>
    <title>This tests that full screen mode with keyboard lock option allows ESC key event.</title>
</head>
<body>
    <div width: 100%; height: 100%;" id="target">
        <p>Test passes if ESC key received by the test case after enter to fullscreen.</p>
    </div>
    <script src="full-screen-test.js"></script>
    <script>
        window.internals.settings.setFullScreenKeyboardLock(true);

        runWithKeyDown(async () => {
            const target = document.querySelector("#target");
            try {
                const keyPressed = (event) => {
                    testExpected("event.key", "Escape");
                };

                await target.requestFullscreen({ keyboardLock: "browser" });
                testExpected("document.fullscreenElement", "[object HTMLDivElement]");
                waitForEvent(
                    document,
                    "keypress",
                    keyPressed,
                    /*endit=*/ false,
                    /*once=*/ true
                );
                eventSender.keyDown("escape");
          } catch (err) {
              logResult(
                  false,
                  `An error occurred while trying to enter full-screen mode: ${err.message}`
              );
          } finally {
              endTest();
          }
        });
    </script>
</body>
</html>
